Fix issues with blank files after save (#523)
* Fix issues with blank files after save

* Add logs

+++ custom_components/hacs/hacsbase/backup.py
@@ -1,10 +1,11 @@
 """Backup."""
 import os
+import shutil
 from time import sleep
-from shutil import copy2, rmtree
-from distutils.dir_util import copy_tree
 
 from integrationhelper import Logger
+
+BACKUP_PATH = "/tmp/hacs_backup/"
 
 
 class Backup:
@@ -14,23 +15,25 @@ class Backup:
         """initialize."""
         self.logger = Logger("hacs.backup")
         self.local_path = local_path
-        self.backup_path = "/tmp/hacs_backup"
+        self.backup_path = f"{BACKUP_PATH}{self.local_path.split('/')[-1]}"
 
     def create(self):
         """Create a backup in /tmp"""
-        if os.path.exists(self.backup_path):
-            rmtree(self.backup_path)
-            while os.path.exists(self.backup_path):
+        if not os.path.exists(self.local_path):
+            return
+        if os.path.exists(BACKUP_PATH):
+            shutil.rmtree(BACKUP_PATH)
+            while os.path.exists(BACKUP_PATH):
                 sleep(0.1)
-        os.makedirs(self.backup_path, exist_ok=True)
+        os.makedirs(BACKUP_PATH, exist_ok=True)
 
         try:
             if os.path.isfile(self.local_path):
-                copy2(self.local_path, self.backup_path)
+                shutil.copyfile(self.local_path, self.backup_path)
                 os.remove(self.local_path)
             else:
-                copy_tree(self.local_path, self.backup_path)
-                rmtree(self.local_path)
+                shutil.copytree(self.local_path, self.backup_path)
+                shutil.rmtree(self.local_path)
                 while os.path.exists(self.local_path):
                     sleep(0.1)
             self.logger.debug(
@@ -41,18 +44,25 @@ class Backup:
 
     def restore(self):
         """Restore from backup."""
+        if not os.path.exists(self.local_path):
+            return
+        if not os.path.exists(self.backup_path):
+            return
+
         if os.path.isfile(self.local_path):
             os.remove(self.local_path)
+            shutil.copyfile(self.backup_path, self.local_path)
         else:
-            rmtree(self.local_path)
+            shutil.rmtree(self.local_path)
             while os.path.exists(self.local_path):
                 sleep(0.1)
-        copy2(self.backup_path, self.local_path)
+            shutil.copytree(self.backup_path, self.local_path)
         self.logger.debug(f"Restored {self.local_path}, from backup {self.backup_path}")
 
     def cleanup(self):
         """Cleanup backup files."""
-        rmtree(self.backup_path)
-        while os.path.exists(self.backup_path):
-            sleep(0.1)
-        self.logger.debug(f"Backup dir {self.backup_path} cleared")
+        if os.path.exists(BACKUP_PATH):
+            shutil.rmtree(BACKUP_PATH)
+            while os.path.exists(BACKUP_PATH):
+                sleep(0.1)
+            self.logger.debug(f"Backup dir {BACKUP_PATH} cleared")
